Update MPolicy to ensure compatibility with MolmoAct server requirements
Modified the state tensor conversion in MPolicy to use float64 instead of float32, aligning with the expectations of some MolmoAct servers that reject float32 payloads. This change enhances the robustness of the policy when interacting with external services by preventing potential 500 errors due to dtype mismatches.

+++ src/lerobot/policies/m/modeling_m.py
@@ -312,7 +312,13 @@ class MPolicy(PreTrainedPolicy):
             state_tensor = batch[OBS_STATE]
             if state_tensor.ndim == 2:
                 state_tensor = state_tensor.squeeze(0)
-            observation["state"] = state_tensor.detach().cpu().float().numpy()
+            # Match ``molmoact.py::send_request`` which forwards
+            # ``obs["joint_positions"]`` as-is (float64 by numpy default).
+            # Some MolmoAct servers are strict about dtype and will reject
+            # float32 payloads with a generic 500 error.
+            observation["state"] = (
+                state_tensor.detach().cpu().double().numpy().astype(np.float64)
+            )
 
         # -- Serialize first so we can log the on-the-wire payload size --
         serialized = json_numpy.dumps(observation)
